util.h: stringb: Add 'const' to the 'static_cast'.
This is needed for compatibility with the libc++ that comes with g++-11,
whereas evidently it was not when using the one with g++-9.  With the
former, and without the cast, usage provokes an error about being unable
to bind a non-const lvalue reference to a temporary.  I don't know why
that error doesn't happen with the g++-9 libc++.